Don't force a directory structure.
Proust uses paths like this:

/memorabilia
/memorabilia/12345/post

wget will encounter the first path and write the response out to a file.
However, when building a directory structure, wget will have no way to
save the latter response, because memorabilia will not be a file.

There's really nothing wget can do about this when building directory
structures, because there's no way to determine that /a and /a/b will
both yield responses.

Solution, then: just write everything out to a flat namespace, letting
wget rename as necessary.  All we care about is that the data eventually
gets into the WARC, anyway, so this isn't really a huge deal.