Include external Javascripts referenced by Proust pages.
These seem to be needed: Proust pages break in funny ways when they're
missing.  Also, systems like IA's Wayback rewrite URLs from e.g.
http://www.example.com/foo.js to
http://wayback/20120106083222js_/http://www.example.com/foo.js, so the
Javascript needs to be present in the WARC for that reference to be
resolved.

The privacy and security implications of this are left as an exercise to
the reader.